+++ HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 1.5.0 - Unreleased
 ------------------
 
+* Add ``get_fingerprints`` to support host keys with all key types.
+  [fschulze]
+
 * Fix jail host ``status`` command for various cases.
   [fschulze]